Move badge display reader to 3dsra_ui
